perf:  use hooks
